Update from February 15, 2022: *) Removed screen dimming *) Simplify process of the rectangle selection *) `RectangleCaptureArea` implements `INotifyPropertyChanged` *) `RectangleCaptureArea` field are in `double` *) Windows is not Topmost

+++ PrintScrn/Services/GraphicsCaptureService.cs
@@ -42,12 +42,12 @@ public class GraphicsCaptureService : IGraphicsCapture
 
     private Bitmap? CaptureBitmapFromScreen(RectangleCaptureArea rectangle)
     {
-        Bitmap? bitmap = new(rectangle.Width, rectangle.Height);
+        Bitmap? bitmap = new((int) rectangle.Width, (int) rectangle.Height);
         var gfx = Graphics.FromImage(bitmap);
         gfx.CompositingQuality = CompositingQuality.HighQuality;
         gfx.SmoothingMode = SmoothingMode.HighQuality;
         gfx.InterpolationMode = InterpolationMode.HighQualityBicubic;
-        gfx.CopyFromScreen(rectangle.X, rectangle.Y, 0, 0, bitmap.Size);
+        gfx.CopyFromScreen((int) rectangle.X, (int) rectangle.Y, 0, 0, bitmap.Size);
         return bitmap;
     }
 }
